// src/__generated__/scriptScheduleDevicesRelay_schedule.graphql.ts
export type scriptScheduleDevicesRelay_schedule$data = {
  readonly assignedDevices: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly " $fragmentSpreads": FragmentRefs<"deviceRowFields_machine">;
      };
    }>;
    readonly filteredCount: number;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly id: string;
  readonly " $fragmentType": "scriptScheduleDevicesRelay_schedule";
};
export type scriptScheduleDevicesRelay_schedule$key = {
  readonly " $data"?: scriptScheduleDevicesRelay_schedule$data;
  readonly " $fragmentSpreads": FragmentRefs<"scriptScheduleDevicesRelay_schedule">;
};